fix(wasm): update TS definitions, add CI check, and satisfy LOC limits
This commit addresses the WASM TypeScript definition drift and CI
failures by:
1. Synchronizing `wasm/chaotic_semantic_memory.d.ts` with the output of
   `wasm-pack build --target web`.
2. Implementing `scripts/check-wasm-freshness.sh` and integrating it
   into the CI workflow to prevent future drift.
3. Refactoring `src/wasm.rs` by moving utility functions and helpers
   to a new `src/wasm_utils.rs` module, ensuring all source files
   stay under the 500 LOC project limit.
4. Removing accidental temporary artifacts (`export.json`).

The updated bindings include all missing methods for GraphRAG,
versioning, and namespace management.

Key changes:
- Added `bfs`, `shortest_path`, `traverse`
- Added `listVersions`, `getVersion`, `rollbackToVersion`
- Added `probe_filtered`, `probe_text`, `inject_text`
- Added `setNamespace`, `getNamespace`
- Refactored WASM modules to satisfy LOC gate
- Fixed CI freshness verification path
