+++ src/lib.rs
@@ -271,12 +271,17 @@ pub mod prelude {
 pub mod wasm;
 #[cfg(target_arch = "wasm32")]
 mod wasm_ext;
-// Include wasm_ext for tests to run underlying data pattern tests on native
+#[cfg(target_arch = "wasm32")]
+mod wasm_graph_rag;
+#[cfg(target_arch = "wasm32")]
+pub mod wasm_utils;
+
+// Include wasm modules for tests to run underlying data pattern tests on native
 #[cfg(all(test, not(target_arch = "wasm32")))]
 mod wasm_ext;
 #[cfg(all(test, not(target_arch = "wasm32")))]
 mod wasm_ext_tests;
-#[cfg(target_arch = "wasm32")]
-mod wasm_graph_rag;
 #[cfg(all(test, not(target_arch = "wasm32")))]
 mod wasm_graph_rag;
+#[cfg(all(test, not(target_arch = "wasm32")))]
+pub mod wasm_utils;
